% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outer.R
\name{util_itc}
\alias{util_itc}
\title{Fit a Binary Intertemporal Choice Model}
\usage{
util_itc(modeltype, choice, amt1, delay1, amt2, delay2)
}
\arguments{
\item{modeltype}{Character. One of "E", "H", "Q", or "GH".
\itemize{
  \item{"E"}: Exponentaial, where U = A * exp(-kD)
  \item{"H"}: Hyperbolic, where U = A/(1+kD)
  \item{"Q"}: Quasi Hyperbolic, where U = A * b * exp(-kD)
  \item{"GH"}: Generalized Hyperbolic, where U = U = A/(1+kD)^s
}}

\item{choice}{Numeric vector of 0s and 1s, where 1 = option 1 chosen.}

\item{amt1, }{prob1 Non-negative Numeric vectors of payoffs (A) and delays (D) for option 1.}

\item{amt2, }{prob2 Non-negative Numeric vectors of payoffs (A) and delays (D) for option 2.}
}
\value{
A list with the following components:
\describe{
  \item{output}{A list with:
    \describe{
      \item{fitted_param}{Estimated model parameter (k, inverse temperature, and optional extra parameter b and s for Quasi Hyperbolic and Generalized Hyperbolic models respectively)}
    }
  }
  \item{fit_metrics}{Limited-memory BFGS}
  \item{modeltype}{The model type used}
  \item{num_observations}{Number of observations}
}
}
\description{
Fits one of four utility-based decision models to binary choice data.
}
\examples{
util_itc("E", c(1,0,0), c(10,10,10), c(1,1,1), c(20,30,40), c(4,5,6))

}
