% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outer.R
\name{util_rc}
\alias{util_rc}
\title{Fit a Utility-Based Risky Choice Model}
\usage{
util_rc(modeltype, choice, amt1, prob1, amt2, prob2)
}
\arguments{
\item{modeltype}{Character. One of "E", "R", "W", or "H".}

\item{choice}{Numeric vector of 0s and 1s, where 1 = option 1 chosen.}

\item{amt1, }{prob1 Numeric vectors of outcomes and probabilities for option 1.}

\item{amt2, }{prob2 Numeric vectors of outcomes and probabilities for option 2.}
}
\value{
A list containing:
\item{params}{Fitted parameter(s) and inverse temperature}
\item{fit_metrics}{Optimization metrics}
\item{modeltype}{The model used}
\item{num_observations}{The number of observations used}
}
\description{
Fits one of several utility-based decision models to binary choice data.
}
\examples{
util_rc("E", c(0,0,1), c(10,10,10), c(1,1,1), c(20,30,40), c(0.6,0.5,0.4))

}
